+++ auth/gauth.go
@@ -1,0 +1,83 @@
+package auth
+
+import (
+	"encoding/json"
+	"fmt"
+	"go-postgres-stocks/models"
+	"net/http"
+	"time"
+
+	"github.com/golang-jwt/jwt/v5"
+)
+
+var secretKey = []byte("secret-key")
+
+func createToken(username string) (string, error) {
+	token := jwt.NewWithClaims(jwt.SigningMethodHS256, jwt.MapClaims{
+		"username": username,
+		"exp":      time.Now().Add(time.Hour * 24).Unix(),
+	})
+
+	tokenString, err := token.SignedString(secretKey)
+	if err != nil {
+		return "", err
+	}
+	return tokenString, nil
+}
+
+func validateToken(tokenString string) error {
+	token, err := jwt.Parse(tokenString, func(token *jwt.Token) (interface{}, error) {
+		return secretKey, nil
+	})
+
+	if err != nil {
+		return err
+	}
+
+	if !token.Valid {
+		return fmt.Errorf("invalid token ")
+	}
+	return nil
+}
+
+func LoginHandler(w http.ResponseWriter, r *http.Request) {
+	w.Header().Set("Content-Type", "Application/json")
+
+	var u models.User
+	json.NewDecoder(r.Body).Decode(&u)
+	fmt.Printf("The request value %v", u)
+
+	if u.Username == "jonSnow" && u.Password == "11110000" {
+		tokenString, err := createToken(u.Username)
+		if err != nil {
+			w.WriteHeader(http.StatusInternalServerError)
+			fmt.Errorf("No username found")
+			return
+		}
+		w.WriteHeader(http.StatusOK)
+		fmt.Fprint(w, tokenString)
+		return
+	} else {
+		w.WriteHeader(http.StatusUnauthorized)
+		fmt.Fprint(w, "Invalid credentials")
+	}
+}
+
+func ProtectedHandler(w http.ResponseWriter, r *http.Request) bool {
+	w.Header().Set("Content-Type", "application-json")
+	tokenString := r.Header.Get("Authorization")
+	if tokenString == "" {
+		w.WriteHeader(http.StatusUnauthorized)
+		fmt.Fprint(w, "Missing authorization header")
+		return false
+	}
+	tokenString = tokenString[len("Bearer "):]
+
+	err := validateToken(tokenString)
+	if err != nil {
+		w.WriteHeader(http.StatusUnauthorized)
+		fmt.Fprint(w, "Invalid token")
+		return false
+	}
+	return true
+}
