+++ middleware/handlers.go
@@ -4,14 +4,13 @@ import (
 	"database/sql"
 	"encoding/json"
 	"fmt"
+	"go-postgres-stocks/auth"
 	"go-postgres-stocks/models"
 	"log"
 	"net/http"
 	"os"
 	"strconv"
-	"time"
 
-	"github.com/golang-jwt/jwt/v5"
 	"github.com/gorilla/mux"
 	"github.com/joho/godotenv"
 	_ "github.com/lib/pq"
@@ -21,8 +20,6 @@ type response struct {
 	ID      int64  `json:"id,omitempty"`
 	Message string `json:"message,omitempty"`
 }
-
-var secretKey = []byte("secret-key")
 
 func createConnection() *sql.DB {
 	err := godotenv.Load(".env")
@@ -49,7 +46,7 @@ func createConnection() *sql.DB {
 }
 
 func CreateStock(w http.ResponseWriter, r *http.Request) {
-	if ProtectedHandler(w, r) {
+	if auth.ProtectedHandler(w, r) {
 		var stock models.Stock
 
 		err := json.NewDecoder(r.Body).Decode(&stock)
@@ -70,7 +67,7 @@ func CreateStock(w http.ResponseWriter, r *http.Request) {
 }
 
 func GetStock(w http.ResponseWriter, r *http.Request) {
-	if ProtectedHandler(w, r) {
+	if auth.ProtectedHandler(w, r) {
 		params := mux.Vars(r)
 
 		id, err := strconv.Atoi(params["id"])
@@ -90,7 +87,7 @@ func GetStock(w http.ResponseWriter, r *http.Request) {
 }
 
 func GetAllStock(w http.ResponseWriter, r *http.Request) {
-	if ProtectedHandler(w, r) {
+	if auth.ProtectedHandler(w, r) {
 		stocks, err := getAllStock()
 
 		if err != nil {
@@ -103,7 +100,7 @@ func GetAllStock(w http.ResponseWriter, r *http.Request) {
 }
 
 func UpdateStock(w http.ResponseWriter, r *http.Request) {
-	if ProtectedHandler(w, r) {
+	if auth.ProtectedHandler(w, r) {
 		params := mux.Vars(r)
 
 		id, err := strconv.Atoi(params["id"])
@@ -133,7 +130,7 @@ func UpdateStock(w http.ResponseWriter, r *http.Request) {
 }
 
 func DeleteStock(w http.ResponseWriter, r *http.Request) {
-	if ProtectedHandler(w, r) {
+	if auth.ProtectedHandler(w, r) {
 		params := mux.Vars(r)
 		id, err := strconv.Atoi(params["id"])
 		if err != nil {
@@ -253,74 +250,4 @@ func deleteStock(id int64) int64 {
 	}
 	fmt.Printf("Total rows/records affected %v", rowsAffected)
 	return rowsAffected
-}
-
-func createToken(username string) (string, error) {
-	token := jwt.NewWithClaims(jwt.SigningMethodHS256, jwt.MapClaims{
-		"username": username,
-		"exp":      time.Now().Add(time.Hour * 24).Unix(),
-	})
-
-	tokenString, err := token.SignedString(secretKey)
-	if err != nil {
-		return "", err
-	}
-	return tokenString, nil
-}
-
-func validateToken(tokenString string) error {
-	token, err := jwt.Parse(tokenString, func(token *jwt.Token) (interface{}, error) {
-		return secretKey, nil
-	})
-
-	if err != nil {
-		return err
-	}
-
-	if !token.Valid {
-		return fmt.Errorf("invalid token ")
-	}
-	return nil
-}
-
-func LoginHandler(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "Application/json")
-
-	var u models.User
-	json.NewDecoder(r.Body).Decode(&u)
-	fmt.Printf("The request value %v", u)
-
-	if u.Username == "jonSnow" && u.Password == "11110000" {
-		tokenString, err := createToken(u.Username)
-		if err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			fmt.Errorf("No username found")
-			return
-		}
-		w.WriteHeader(http.StatusOK)
-		fmt.Fprint(w, tokenString)
-		return
-	} else {
-		w.WriteHeader(http.StatusUnauthorized)
-		fmt.Fprint(w, "Invalid credentials")
-	}
-}
-
-func ProtectedHandler(w http.ResponseWriter, r *http.Request) bool {
-	w.Header().Set("Content-Type", "application-json")
-	tokenString := r.Header.Get("Authorization")
-	if tokenString == "" {
-		w.WriteHeader(http.StatusUnauthorized)
-		fmt.Fprint(w, "Missing authorization header")
-		return false
-	}
-	tokenString = tokenString[len("Bearer "):]
-
-	err := validateToken(tokenString)
-	if err != nil {
-		w.WriteHeader(http.StatusUnauthorized)
-		fmt.Fprint(w, "Invalid token")
-		return false
-	}
-	return true
 }
