+++ router/router.go
@@ -1,6 +1,7 @@
 package router
 
 import (
+	"go-postgres-stocks/auth"
 	"go-postgres-stocks/middleware"
 
 	"github.com/gorilla/mux"
@@ -14,7 +15,7 @@ func Router() *mux.Router {
 	router.HandleFunc("/api/newstock", middleware.CreateStock).Methods("POST", "OPTIONS")
 	router.HandleFunc("/api/stock/{id}", middleware.UpdateStock).Methods("PUT", "OPTIONS")
 	router.HandleFunc("/api/deletestock/{id}", middleware.DeleteStock).Methods("DELETE", "OPTIONS")
-	router.HandleFunc("/api/login", middleware.LoginHandler).Methods("POST", "OPTIONS")
+	router.HandleFunc("/api/login", auth.LoginHandler).Methods("POST", "OPTIONS")
 
 	return router
 }
